+++ source/iterator/snapshot.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/conduitio-labs/conduit-connector-oracle/coltypes"
+	"github.com/conduitio-labs/conduit-connector-oracle/columntypes"
 	"github.com/conduitio-labs/conduit-connector-oracle/repository"
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	"github.com/jmoiron/sqlx"
@@ -48,7 +48,7 @@ type Snapshot struct {
 
 	rows *sqlx.Rows
 	// columnTypes represents a columns' data from table
-	columnTypes map[string]coltypes.ColumnData
+	columnTypes map[string]columntypes.ColumnData
 }
 
 // SnapshotParams represents an incoming params for the NewSnapshot function.
@@ -61,7 +61,7 @@ type SnapshotParams struct {
 	OrderingColumn string
 	Columns        []string
 	BatchSize      int
-	ColumnTypes    map[string]coltypes.ColumnData
+	ColumnTypes    map[string]columntypes.ColumnData
 }
 
 // NewSnapshot creates a new instance of the Snapshot iterator.
@@ -106,7 +106,7 @@ func (i *Snapshot) Next(_ context.Context) (sdk.Record, error) {
 		return sdk.Record{}, fmt.Errorf("scan rows: %w", err)
 	}
 
-	transformedRow, err := coltypes.TransformRow(row, i.columnTypes)
+	transformedRow, err := columntypes.TransformRow(row, i.columnTypes)
 	if err != nil {
 		return sdk.Record{}, fmt.Errorf("transform row column types: %w", err)
 	}
